chore(packetio): improve packetio
This commit improves the packet io slightly by avoiding the packet
creation work if there is no event listener for the packet which is
about to be created. The side effect of this is that the 'packet' event
has been removed. The data handling method has also been improved
slightly by using buffer copying instead of a for loop.